Refactor :
Moved chart data and options getter from InventoryStore to ReportsView. Since it's only used there
