+++ src/stores/inventoryStore.js
@@ -958,89 +958,6 @@ export const useInventoryStore = defineStore('inventory', {
 
     /**
      * @async
-     * @method getCategoryChartData
-     * @returns {Object} Chart data for categories
-     * @description Gets chart data for categories
-     */
-    getCategoryChartData() {
-      // Group items by category and calculate total sales
-      const categoryData = this.items.reduce((acc, item) => {
-        if (!acc[item.category])
-          acc[item.category] = 0
-        // Assuming each item's sales contribution is quantity * price
-        acc[item.category] += item.quantity * item.price
-        return acc
-      }, {})
-      const labels = Object.keys(categoryData).map(category =>
-        category.charAt(0).toUpperCase() + category.slice(1)
-      )
-      const data = Object.values(categoryData)
-
-      const colors = [
-        '#FF6384',
-        '#36A2EB',
-        '#FFCE56',
-        '#4BC0C0',
-        '#9966FF',
-        '#FF9F40'
-      ]
-
-      return {
-        labels,
-        datasets: [{
-          data,
-          backgroundColor: colors.slice(0, labels.length),
-          hoverBackgroundColor: colors.slice(0, labels.length)
-        }]
-      }
-    },
-
-    /**
-     * @async
-     * @method getCategoryChartOptions
-     * @param {string} textColor - Text color for the chart
-     * @returns {Object} Chart options for categories
-     * @description Gets chart options for categories
-     */
-    getCategoryChartOptions(textColor) {
-      return {
-        responsive: true,
-        maintainAspectRatio: false,
-        plugins: {
-          legend: {
-            position: 'right',
-            labels: {
-              color: textColor,
-              font: {
-                size: 12
-              }
-            }
-          },
-          title: {
-            display: true,
-            text: 'Sales by Category',
-            color: textColor,
-            font: {
-              size: 16
-            }
-          },
-          tooltip: {
-            callbacks: {
-              label: (context) => {
-                const label = context.label || ''
-                const value = context.raw || 0
-                const total = context.dataset.data.reduce((a, b) => a + b, 0)
-                const percentage = ((value / total) * 100).toFixed(1)
-                return `${label}: ${financialStore.formatCurrency(value)} (${percentage}%)`
-              }
-            }
-          }
-        }
-      }
-    },
-
-    /**
-     * @async
      * @method openItemDialog
      * @param {Object|null} item - Item to edit or null for new item
      * @returns {void}
